Refactoriza comando de traducción de nombre de permisos
